add require flag to mocha options within profile
This PR adds a `"require"` property that can be a string (name of an npm module) or an array (of npm module names). The npm modules defined here are then `require`d before we instantiate the `Mocha` instances.

If a `require` fails, we throw the error

also removed gulp, which was only running eslint and there was duplicate eslint config in `.eslintrc.js`