prevent all unnecessary mocha.run()
When things are run with `"parallel": "file"`, instances of mocha are
created and run that do not need to be (all files from the `baseDir`
are sent as an instance before a `grep` is run on them).

This PR checks to see if it's "parallel by file" and then that there are
actually grep matches on that file before `mocha.run()` is invoked.